Provide default implementations for Idiom's map and Monad's app
